refactor: replace DraftVariableEntity with common entity
