+++ tasklist/webapp/src/main/java/io/camunda/tasklist/webapp/api/rest/v1/entities/VariableResponse.java
@@ -7,9 +7,9 @@
  */
 package io.camunda.tasklist.webapp.api.rest.v1.entities;
 
-import io.camunda.tasklist.entities.DraftTaskVariableEntity;
 import io.camunda.tasklist.entities.TaskVariableEntity;
 import io.camunda.tasklist.entities.VariableEntity;
+import io.camunda.webapps.schema.entities.tasklist.DraftTaskVariableEntity;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.Objects;
 import java.util.StringJoiner;
