+++ webapps-schema/src/main/java/io/camunda/webapps/schema/entities/tasklist/DraftTaskVariableEntity.java
@@ -5,24 +5,18 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.tasklist.entities;
+package io.camunda.webapps.schema.entities.tasklist;
 
 import java.util.Objects;
 
 /** Represents draft variable with its value when task is in created state. */
-public class DraftTaskVariableEntity extends TasklistZeebeEntity<DraftTaskVariableEntity> {
+public class DraftTaskVariableEntity extends TasklistEntity<DraftTaskVariableEntity> {
 
   private String taskId;
   private String name;
   private String value;
   private String fullValue;
   private boolean isPreview;
-
-  public DraftTaskVariableEntity() {}
-
-  public static String getIdBy(String idPrefix, String name) {
-    return String.format("%s-%s", idPrefix, name);
-  }
 
   public String getTaskId() {
     return taskId;
@@ -69,47 +63,9 @@ public class DraftTaskVariableEntity extends TasklistZeebeEntity<DraftTaskVariab
     return this;
   }
 
-  public static DraftTaskVariableEntity createFrom(
-      TaskEntity taskEntity, String name, String value, int variableSizeThreshold) {
-    return completeVariableSetup(
-        new DraftTaskVariableEntity().setId(getIdBy(taskEntity.getId(), name)),
-        taskEntity,
-        name,
-        value,
-        variableSizeThreshold);
-  }
-
-  public static DraftTaskVariableEntity createFrom(
-      String draftVariableId,
-      TaskEntity taskEntity,
-      String name,
-      String value,
-      int variableSizeThreshold) {
-    return completeVariableSetup(
-        new DraftTaskVariableEntity().setId(draftVariableId),
-        taskEntity,
-        name,
-        value,
-        variableSizeThreshold);
-  }
-
-  private static DraftTaskVariableEntity completeVariableSetup(
-      DraftTaskVariableEntity entity,
-      TaskEntity taskEntity,
-      String name,
-      String value,
-      int variableSizeThreshold) {
-
-    entity.setTaskId(taskEntity.getId()).setName(name).setTenantId(taskEntity.getTenantId());
-    if (value.length() > variableSizeThreshold) {
-      // store preview
-      entity.setValue(value.substring(0, variableSizeThreshold));
-      entity.setIsPreview(true);
-    } else {
-      entity.setValue(value);
-    }
-    entity.setFullValue(value);
-    return entity;
+  @Override
+  public int hashCode() {
+    return Objects.hash(super.hashCode(), taskId, name, value, fullValue, isPreview);
   }
 
   @Override
@@ -129,10 +85,5 @@ public class DraftTaskVariableEntity extends TasklistZeebeEntity<DraftTaskVariab
         && Objects.equals(name, that.name)
         && Objects.equals(value, that.value)
         && Objects.equals(fullValue, that.fullValue);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(super.hashCode(), taskId, name, value, fullValue, isPreview);
   }
 }
